ALSA: info - Use standard types for info callbacks
Use loff_t, size_t and ssize_t for arguments of info callbacks
to follow the standard procfs.

+++ sound/pci/mixart/mixart.c
@@ -1103,11 +1103,10 @@ static int snd_mixart_free(struct mixart_mgr *mgr)
 /*
  * proc interface
  */
-static long long snd_mixart_BA0_llseek(struct snd_info_entry *entry,
-				       void *private_file_data,
-				       struct file *file,
-				       long long offset,
-				       int orig)
+static loff_t snd_mixart_BA0_llseek(struct snd_info_entry *entry,
+				    void *private_file_data,
+				    struct file *file,
+				    loff_t offset, int orig)
 {
 	offset = offset & ~3; /* 4 bytes aligned */
 
@@ -1129,11 +1128,10 @@ static long long snd_mixart_BA0_llseek(struct snd_info_entry *entry,
 	return file->f_pos;
 }
 
-static long long snd_mixart_BA1_llseek(struct snd_info_entry *entry,
-				       void *private_file_data,
-				       struct file *file,
-				       long long offset,
-				       int orig)
+static loff_t snd_mixart_BA1_llseek(struct snd_info_entry *entry,
+				    void *private_file_data,
+				    struct file *file,
+				    loff_t offset, int orig)
 {
 	offset = offset & ~3; /* 4 bytes aligned */
 
@@ -1158,9 +1156,10 @@ static long long snd_mixart_BA1_llseek(struct snd_info_entry *entry,
 /*
   mixart_BA0 proc interface for BAR 0 - read callback
  */
-static long snd_mixart_BA0_read(struct snd_info_entry *entry, void *file_private_data,
-				struct file *file, char __user *buf,
-				unsigned long count, unsigned long pos)
+static ssize_t snd_mixart_BA0_read(struct snd_info_entry *entry,
+				   void *file_private_data,
+				   struct file *file, char __user *buf,
+				   size_t count, loff_t pos)
 {
 	struct mixart_mgr *mgr = entry->private_data;
 	unsigned long maxsize;
@@ -1179,9 +1178,10 @@ static long snd_mixart_BA0_read(struct snd_info_entry *entry, void *file_private
 /*
   mixart_BA1 proc interface for BAR 1 - read callback
  */
-static long snd_mixart_BA1_read(struct snd_info_entry *entry, void *file_private_data,
-				struct file *file, char __user *buf,
-				unsigned long count, unsigned long pos)
+static ssize_t snd_mixart_BA1_read(struct snd_info_entry *entry,
+				   void *file_private_data,
+				   struct file *file, char __user *buf,
+				   size_t count, loff_t pos)
 {
 	struct mixart_mgr *mgr = entry->private_data;
 	unsigned long maxsize;
